Change from Baseline Plot Teal Module

This module produces a plot of the mean (of value at visit or change from baseline) by arm, that's similar to STREAM template MNG01

tm_chgfbl_plot(label, paramcd, paramcd_choices = paramcd, arm_var = "ARM",
  arm_var_choices = arm_var, ytype = "CHG", ytype_choices = ytype,
  errbar = "SE", errbar_choices = errbar, ref_line_txt = NULL,
  plot_height = c(800, 400, 2000), font_size = c(20, 10, 30),
  n_rotate = TRUE,
  pre_output = helpText("Note that currently missing values are omitted. Grouping legend can be found in the table"),
  post_output = NULL)

Arguments

label
full name label of module

paramcd
filter the rows in AQS given the paramcd value

paramcd_choices
choices of possible paramcd

arm_var
selected variable to use as arms

arm_var_choices
choices of arm variables

ytype
selected type of value to plot for y-axis

ytype_choices
choices of possible ytype: AVAL=value at visit, CHG = change from baseline, PCHG = % change from baseline

errbar
selected type of error bar for the plot

errbar_choices
choices of types of error bar: SE = standarnd error, SD = standard deviation, 95CI = 95% confidence interval of the mean

ref_line_txt
horizontal reference lines to display on plot, entered as text separated by comma

plot_height
height of the change from baseline plot

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

arm_label
label to display on plot for the selected arm variable

font_sie
fontsize to pass to plotting function

Value

an module object

Details

Package forcats used to re-format arm data into leveled factors.

Examples

 

# NOT RUN {
library(atezo.data)
library(dplyr)
library(tidyr)
library(teal.oncology)
library(forcats)
library(ggplot2)
library(grid)
library(gridExtra)

ASL <- asl(com.roche.cdpt7722.wo29637.rl)
AQS <- aqs(com.roche.cdpt7722.wo29637.rl)

options(teal_logging = FALSE)

parammap <- unique(AQS[c("PARAM", "PARAMCD")])
paramcd_choices_labelled <- setNames(parammap$PARAMCD, paste(parammap$PARAMCD, parammap$PARAM, sep = " - "))

arm_var_choices_list <- c("ARM", "ARMCD", "AGE65", "SEX", "HSTSTYP", "PMETA",
                          "IC","ICC", "TCC", "ICCAT1", "ICCAT2", "ICCAT3", "BIO2ICL")
arm_var_choices_labelled <- setNames(arm_var_choices_list,
                            paste(arm_var_choices_list, labels_over_names(ASL[arm_var_choices_list]), sep=" - "))

x <- teal::init(
  data = list(ASL = ASL, AQS = AQS),
  modules = root_modules(
    #tm_variable_browser(),
    #tm_data_table(),
    tm_chgfbl_plot(
       label = "Change from Baseline Plot",
       paramcd = "FATIGI",
       paramcd_choices = paramcd_choices_labelled,
       arm_var = "ARM",
       arm_var_choices = arm_var_choices_labelled,
       ytype = "CHG",
       ytype_choices = c("CHG", "AVAL"),
       errbar = "SE",
       errbar_choices = c("SE", "SD", "95CI", "IQR")
   ))
)
shinyApp(x$ui, x$server)

# }