Forest Response Plot teal module

Forest Response Plot teal module

tm_forest_response(label, paramcd = "OVRSPI", paramcd_choices = paramcd,
  arm_var = "ARM", arm_var_choices = arm_var, subgroup_var,
  subgroup_var_choices = subgroup_var, plot_height = c(700, 200, 2000),
  cex = 1.3,
  pre_output = helpText("graph needs to be of a certain width to be displayed"),
  post_output = NULL)

Arguments

label
a character string displayed as module label

paramcd
default response type from PARAMCD

paramcd_choices
a vector of possible paramcd

arm_var
default variable name used as the arm variable

arm_var_choices
a character vector for the choices of arm_var

subgroup_var
a vector of variable names used as the default subgroups

subgroup_var_choices
a vector of variable names to choose the subgroup_var from

plot_height
height of the forest plot

cex
multiplier applied to overall fontsize

pre_output
text displayed at the top of the plot

post_output
text displayed at the bottom of the plot

Examples

  

# NOT RUN {
library(teal.oncology)
library(atezo.data)
library(dplyr)
library(forcats)

ARS <- ars(com.roche.cdt30019.go29436.re)
ASL <- asl(com.roche.cdt30019.go29436.re)

ASL$BAGED <- ifelse(ASL$BAGE <= median(ASL$BAGE), "<=median", ">median")

x <- teal::init(
  data = list(ASL = ASL, ARS = ARS),
  modules = root_modules(
    tm_forest_response(
       label = "Forest Response",
       paramcd = "OVRSPI",
       paramcd_choices = c("BESRSPI","LSTASDI","MBESRSPI","MLSTASDI","OVRSPI"),
       plot_height = c(600, 200, 2000),
       subgroup_var = c("BAGED", "SEX", "BECOG"),
       subgroup_var_choices = names(ASL),
       arm_var = "ARM",
       arm_var_choices = c("ARM", "ARMCD", "ACTARM")
   )
  )
)
shinyApp(x$ui, x$server)

# }